Update gpu_checker_api.py
Made changes to how we identify if a product is in stock - instead of relying on isAvailable we've now changed is so that we monitor for SKU changes for each monitored product. If the SKU changes the product release is probably imminent so we then start checking the inventory API status if is_active changes to true and in that case send out the purchaseLink for the product that has come into stock.